Move debug functionality to local development environment
Consolidate includes/debug.php and includes/class-debug.php into a single
local/class-debug.php under the Development namespace, so all debug tooling
(admin UI visibility for post types/taxonomies and activity logging) only
loads in local environments.
