+++ activitypub.php
@@ -128,14 +128,6 @@ function plugin_init() {
 			require_once $loader_file;
 		}
 	}
-
-	if ( \defined( 'WP_DEBUG' ) && WP_DEBUG ) {
-		$debug_file = __DIR__ . '/includes/debug.php';
-		if ( \file_exists( $debug_file ) && \is_readable( $debug_file ) ) {
-			require_once $debug_file;
-			Debug::init();
-		}
-	}
 }
 \add_action( 'plugins_loaded', __NAMESPACE__ . '\plugin_init' );
